+++ spec/rack_jquery_spec.rb
@@ -3,22 +3,30 @@
 require 'spec_helper'
 require_relative "../lib/rack/backbone.rb"
 
+class Rack::Backbone # for clarity!
+
 describe "The class methods" do
-  let(:env) { {} }
-  subject { Rack::Backbone.cdn env, :organisation => organisation }
+  let(:path) {::File.join(DEFAULT_OPTIONS[:http_path],BACKBONE_FILE_NAME)}
+  let(:env) { {"rack.backbone.http_path" => path} }
+  let(:default_options) { {} }
+  subject(:cdn) { Rack::Backbone.cdn env, default_options.merge(options) }
 
   context "Given the organisation option" do
     context "of nil (the default)" do
-      let(:organisation) { nil }
-      it { should == "<script src='#{Rack::Backbone::CDN::CLOUDFLARE}'></script>\n#{Rack::Backbone::FALLBACK}" }
+      let(:options) { {:organisation => nil } }
+      it { should == "<script src='#{CDN::CLOUDFLARE}'></script>\n#{FALLBACK_TOP}#{path}#{FALLBACK_BOTTOM}" }
     end
     context "of :jsdelivr" do
-      let(:organisation) { :jsdelivr }
-      it { should == "<script src='#{Rack::Backbone::CDN::JSDELIVR}'></script>\n#{Rack::Backbone::FALLBACK}" }
+      let(:options) { {:organisation => :jsdelivr } }
+      it { should == "<script src='#{CDN::JSDELIVR}'></script>\n#{FALLBACK_TOP}#{path}#{FALLBACK_BOTTOM}" }
     end
     context "of :cloudflare" do
-      let(:organisation) { :cloudflare }
-      it { should == "<script src='#{Rack::Backbone::CDN::CLOUDFLARE}'></script>\n#{Rack::Backbone::FALLBACK}" }
+      let(:options) { {:organisation => :cloudflare } }
+      it { should == "<script src='#{CDN::CLOUDFLARE}'></script>\n#{FALLBACK_TOP}#{path}#{FALLBACK_BOTTOM}" }
+    end
+    context "of false, to get the fallback script only" do
+      let(:options) { {:organisation => false } }
+      it { should == "<script src='#{path}'></script>" }    
     end
   end
 
@@ -56,7 +64,7 @@ describe "Inserting the CDN" do
       end
       it_should_behave_like "Any route"
       subject { last_response.body }
-      let(:expected) { Rack::Backbone::CDN::JSDELIVR }
+      let(:expected) { CDN::JSDELIVR }
       it { should include expected }
     end
     context "Unspecified CDN" do
@@ -65,7 +73,7 @@ describe "Inserting the CDN" do
       end
       it_should_behave_like "Any route"
       subject { last_response.body }
-      let(:expected) { Rack::Backbone::CDN::CLOUDFLARE }
+      let(:expected) { CDN::CLOUDFLARE }
       it { should include expected }
     end
     context "Cloudflare CDN" do
@@ -74,7 +82,7 @@ describe "Inserting the CDN" do
       end
       it_should_behave_like "Any route"
       subject { last_response.body }
-      let(:expected) { Rack::Backbone::CDN::CLOUDFLARE }
+      let(:expected) { CDN::CLOUDFLARE }
       it { should include expected }
     end
   end
@@ -92,7 +100,7 @@ describe "Inserting the CDN" do
       end
       it_should_behave_like "Any route"
       subject { last_response.body }
-      let(:expected) { Rack::Backbone::CDN::JSDELIVR }
+      let(:expected) { CDN::JSDELIVR }
       it { should include expected }
     end
     context "Unspecified CDN" do
@@ -101,7 +109,7 @@ describe "Inserting the CDN" do
       end
       it_should_behave_like "Any route"
       subject { last_response.body }
-      let(:expected) { Rack::Backbone::CDN::CLOUDFLARE }
+      let(:expected) { CDN::CLOUDFLARE }
       it { should include expected }
     end
     context "Cloudflare CDN" do
@@ -110,7 +118,7 @@ describe "Inserting the CDN" do
       end
       it_should_behave_like "Any route"
       subject { last_response.body }
-      let(:expected) { Rack::Backbone::CDN::CLOUDFLARE }
+      let(:expected) { CDN::CLOUDFLARE }
       it { should include expected }
     end
   end
@@ -121,24 +129,53 @@ require 'timecop'
 require 'time'
 
 describe "Serving the fallback backbone" do
-  include_context "All routes"
   before do
-    get "/js/backbone-#{Rack::Backbone::BACKBONE_VERSION}-min.js"
+    get path
   end
-  it_should_behave_like "Any route"
   subject { last_response.body }
-  it { should start_with "(function(){var t=this;var e=t.Backbone;" }
+  let(:path){ ::File.join(http_path, BACKBONE_FILE_NAME) }
 
-  context "Re requests" do
-    before do
-      at_start = Time.parse(Rack::Backbone::BACKBONE_VERSION_DATE) + 60 * 60 * 24 * 180
-      Timecop.freeze at_start
-      get "/js/backbone-#{Rack::Backbone::BACKBONE_VERSION}-min.js"
-      Timecop.travel Time.now + 86400 # add a day
-      get "/js/backbone-#{Rack::Backbone::BACKBONE_VERSION}-min.js", {}, {"HTTP_IF_MODIFIED_SINCE" => Rack::Utils.rfc2109(at_start) }
+  context "With the default :http_path (none given)" do
+    include_context "All routes"
+    let(:http_path) { DEFAULT_OPTIONS[:http_path] }
+
+    it_should_behave_like "Any route"
+    it { should start_with "(function(){var t=this;var e=t.Backbone;" }
+
+    context "Re requests" do
+      before do
+        at_start = Time.parse(BACKBONE_VERSION_DATE) + 60 * 60 * 24 * 180
+        Timecop.freeze at_start
+        get path
+        Timecop.travel Time.now + 86400 # add a day
+        get path, {}, {"HTTP_IF_MODIFIED_SINCE" => Rack::Utils.rfc2109(at_start) }
+      end
+      subject { last_response }
+      its(:status) { should == 304 }
     end
-    subject { last_response }
-    its(:status) { should == 304 }
-    
   end
+  context "Given a different http_path via the options" do
+    include_context "All routes" do
+      let(:app) {
+        Sinatra.new do
+          use Rack::JQuery
+          use Rack::Lodash
+          use Rack::Backbone, :http_path => "/assets/javascripts"
+        end
+      }
+    end
+    context "That is valid" do
+      let(:http_path) { "/assets/javascripts" }
+
+      it_should_behave_like "Any route"
+      it { should start_with "(function(){var t=this;var e=t.Backbone;" }
+    end
+    context "That is not valid" do
+      let(:http_path) { "/this/is/not/the/path/it/was/setup/with" }
+      subject { last_response }
+      it { should_not be_ok }
+    end
+  end
+end
+
 end