+++ CHANGES.md
@@ -1,5 +1,13 @@
 # CH CH CH CHANGES #
 
+## Next release ##
+
+* Can pass `false` to the `cdn` method to just get the fallback with no CDN, useful for working locally.
+* Fixed problem with the :http_path option for the fallback route.
+
+----
+
+
 ## Thursday the 19th of September 2013 ##
 
 ### v0.0.4 ###